Bad design. Determined to reapproach with cleaner solution employing builders.
Creating diverse object from abstract DA class proved infeasible consequently that composite objects (Department, Item, Admin) required inputs from proxy DA objects- and the proxies needed the parsed object of the composites.
The double reliance was not infeasible, but so tightly coupled and complicated that it was determine it would be better to create internal builders.